Splitting up lineshapes
