+++ include/goofit/PDFs/physics/lineshapes/RBW.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <goofit/PDFs/physics/lineshapes/Lineshape.h>
+
+namespace GooFit {
+namespace Lineshapes {
+
+class RBW : public Lineshape {
+  public:
+    RBW(std::string name,
+        Variable mass,
+        Variable width,
+        unsigned int L,
+        unsigned int Mpair,
+        FF FormFac    = FF::BL_Prime,
+        fptype radius = 1.5);
+
+    ~RBW() override = default;
+};
+
+} // namespace Lineshapes
+} // namespace GooFit
